dom/media/test/test_hls_player_independency.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test playback of 2 HLS video at the same page </title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
  <script type="text/javascript" src="manifest.js"></script>
</head>
<body>
    <div id='player1'>
      <video id='player4x3' controls autoplay>
      </video>
    </div>
    <p> 4x3 basic stream<span>
    <span>
    <div height = 10>
    <span>
    <div id='player2'>
      <video id='player16x9' controls autoplay>
      </video>
    </div>
    <p> 16x9 basic stream<span>

<script class="testbody" type="text/javascript">

function startTest() {
  var v4x3 = document.getElementById('player4x3');
  var v16x9 = document.getElementById('player16x9');

  var p1 = once(v4x3, 'ended', function onended(e) {
    is(v4x3.videoWidth, 400, "4x3 content, the width should be 400.");
    is(v4x3.videoHeight, 300, "4x3 content, the height should be 300.");
  });

  var p2 = once(v16x9, 'ended', function onended(e) {
    is(v16x9.videoWidth, 416, "16x9 content, the width should be 416.");
    is(v16x9.videoHeight, 234, "16x9 content, the height should be 234.");
  });

  v4x3.src = serverUrl + "/bipbop_4x3_single.m3u8";
  v16x9.src = serverUrl + "/bipbop_16x9_single.m3u8";
  Promise.all([p1, p2]).then(() => {
    SimpleTest.finish();
  });
}

startTest();
SimpleTest.waitForExplicitFinish();

</script>
</pre>
</body>
</html>
